Rename CloudTestDataConfig to ImpressionTestDataConfig
Also removes mutable populationSpecResourcePath field by parsing
config directly in run().
