+++ src/main/kotlin/org/wfanet/measurement/integration/common/BUILD.bazel
@@ -349,7 +349,7 @@ kt_jvm_library(
         ":event_group_config",
         "//src/main/kotlin/org/wfanet/measurement/loadtest/dataprovider:synthetic_data_generation",
         "//src/main/proto/wfa/measurement/api/v2alpha/event_group_metadata/testing:simulator_synthetic_data_spec_kt_jvm_proto",
-        "//src/main/proto/wfa/measurement/integration/k8s/testing:cloud_test_data_config_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/integration/k8s/testing:impression_test_data_config_kt_jvm_proto",
     ],
 )
