+++ src/main/proto/wfa/measurement/integration/k8s/testing/BUILD.bazel
@@ -37,8 +37,8 @@ kt_jvm_proto_library(
 )
 
 proto_library(
-    name = "cloud_test_data_config_proto",
-    srcs = ["cloud_test_data_config.proto"],
+    name = "impression_test_data_config_proto",
+    srcs = ["impression_test_data_config.proto"],
     strip_import_prefix = IMPORT_PREFIX,
     deps = [
         "@com_google_protobuf//:struct_proto",
@@ -46,12 +46,12 @@ proto_library(
 )
 
 kt_jvm_proto_library(
-    name = "cloud_test_data_config_kt_jvm_proto",
+    name = "impression_test_data_config_kt_jvm_proto",
     visibility = [
         "//src/main/kotlin/org/wfanet/measurement/integration/common:__pkg__",
         "//src/main/kotlin/org/wfanet/measurement/loadtest/edpaggregator/tools:__pkg__",
         "//src/test/kotlin/org/wfanet/measurement/integration/k8s:__pkg__",
         "//src/test/kotlin/org/wfanet/measurement/integration/common:__pkg__",
     ],
-    deps = [":cloud_test_data_config_proto"],
+    deps = [":impression_test_data_config_proto"],
 )
