+++ src/test/kotlin/org/wfanet/measurement/integration/common/BUILD.bazel
@@ -41,7 +41,7 @@ kt_jvm_test(
         "//src/main/kotlin/org/wfanet/measurement/integration/common:event_group_config",
         "//src/main/kotlin/org/wfanet/measurement/loadtest/edpaggregator/tools:generate_synthetic_data",
         "//src/main/proto/wfa/measurement/api/v2alpha/event_group_metadata/testing:simulator_synthetic_data_spec_kt_jvm_proto",
-        "//src/main/proto/wfa/measurement/integration/k8s/testing:cloud_test_data_config_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/integration/k8s/testing:impression_test_data_config_kt_jvm_proto",
         "@com_google_protobuf//:protobuf_java",
         "@wfa_common_jvm//imports/java/com/google/common/truth",
         "@wfa_common_jvm//imports/java/org/junit",
