+++ src/test/kotlin/org/wfanet/measurement/integration/k8s/BUILD.bazel
@@ -207,7 +207,7 @@ kt_jvm_library(
         "//src/main/kotlin/org/wfanet/measurement/integration/common:event_group_config",
         "//src/main/proto/wfa/measurement/api/v2alpha/event_group_metadata/testing:simulator_synthetic_data_spec_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/api/v2alpha/event_templates/testing:test_event_kt_jvm_proto",
-        "//src/main/proto/wfa/measurement/integration/k8s/testing:cloud_test_data_config_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/integration/k8s/testing:impression_test_data_config_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/integration/k8s/testing:edpa_correctness_test_config_kt_jvm_proto",
         "@wfa_common_jvm//imports/java/org/junit",
         "@wfa_common_jvm//imports/kotlin/kotlinx/coroutines:core",
@@ -222,7 +222,7 @@ java_test(
     name = "EdpAggregatorCorrectnessTest",
     timeout = "long",
     data = [
-        ":cloud_test_data_config.textproto",
+        ":impression_test_data_config.textproto",
         ":edpa_correctness_test_config.textproto",
     ],
     tags = [
